+++ docs/changelog.rst
@@ -1,6 +1,7 @@
 Next version (not yet released)
 -------------------------------
 
+- Add pip_constraint entry to config yaml file
 
 
 Grocker 2.4.2 (2016-04-11)